+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+0.3.2 Changelog
+====
+
+Breaking Changes
+---
+
+- case field is null all converters and validators for this field will be ignore.
+- no more needed to handle nulls in from/to converters.
+
 0.3.1 Changelog
 ====